requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.4.0
backports.zoneinfo==0.2.1
colorama==0.4.4
Django==4.0.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
iniconfig==1.1.1
packaging==21.3
Pillow==9.0.0
pluggy==1.0.0
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
